Convert physical_volume to a custom resource
